Make gensim/archsim compile with LLVM
Changes include:

* Add missing header files
* Add declaration of functions operator<< to the data structures
  that need it.
  Move the definition of the operator<< for Address into its own .cpp file
* Use the termios struct rather than termio for ioctl.
* To silence some warning add the override keyword to member functions.
* In gensim, make sure not to repeat the default template initializer
  in the declaration of helper functions.
  The function declaration:
  template <bool trace=false> XXX();
  should be matched by the function defintion:
  template <bool trace> XXX() { ... }
  not
  template <bool trace=false> XXX() { ... }
  Repeating the default template initializer makes llvm trigger an error.


Make StateBlockDescriptor::AddBlock return the member variable total_size_

The function StateBlockDescriptor::AddBlock was marked as returning uint32_t
but no return was present in the body.
This was causing LLVM to insert a ud2 instruction (undefined) triggering a runtime failure.


Use tabs instead of spaces


Fix two typos


Merge


* Add cmake option ARCHSIM_DISABLE_VALGRIND to control the use of valgrind


Fix a ton of warnings


Fix tabs


Merge from default

+++ archsim/src/blockjit/IRPrinter.cpp
@@ -1,8 +1,4 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
+/* This file is Copyright University of Edinburgh 2018. For license details, see LICENSE. */
 
 #include "blockjit/IRPrinter.h"
 #include <iomanip>
@@ -18,7 +14,7 @@ static void dump_insn(const IRInstruction *insn, std::ostream &str)
 
 	str << " " << std::left << std::setw(12) << std::setfill(' ') << descr->mnemonic;
 
-	for (int op_idx = 0; op_idx < insn->operands.size(); op_idx++) {
+	for (size_t op_idx = 0; op_idx < insn->operands.size(); op_idx++) {
 		const IROperand *oper = &insn->operands[op_idx];
 
 		if (descr->format[op_idx] != 'X') {
@@ -45,7 +41,8 @@ static void dump_insn(const IRInstruction *insn, std::ostream &str)
 }
 
 
-void IRPrinter::DumpIR(std::ostream& ostr, const captive::arch::jit::TranslationContext& ctx) {
+void IRPrinter::DumpIR(std::ostream& ostr, const captive::arch::jit::TranslationContext& ctx)
+{
 	IRBlockId current_block_id = INVALID_BLOCK_ID;
 
 	for (uint32_t ir_idx = 0; ir_idx < ctx.count(); ir_idx++) {
